pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "3DFeatProbe"
version = "0.1.0"
description = "Analyzing 3D awareness in Vision Transformers"
requires-python = ">=3.11"
dependencies = [
    "addict==2.4.0",
    "antlr4-python3-runtime==4.9.3",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "av==12.0.0",
    "blinker==1.9.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "comm==0.2.3",
    "ConfigArgParse==1.7.1",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dash==3.2.0",
    "decorator==5.2.1",
    "einops==0.8.1",
    "executing==2.2.1",
    "fastjsonschema==2.21.2",
    "filelock==3.19.1",
    "Flask==3.1.2",
    "fonttools==4.60.0",
    "fsspec==2025.9.0",
    "fvcore==0.1.5.post20221221",
    "greenlet==3.2.4",
    "h5py==3.14.0",
    "hf-xet==1.1.10",
    "huggingface-hub==0.35.0",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib_metadata==8.7.0",
    "ipython==9.6.0",
    "ipython_pygments_lexers==1.1.1",
    "ipywidgets==8.1.7",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter_core==5.8.1",
    "jupyterlab_widgets==3.0.15",
    "kiwisolver==1.4.9",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "narwhals==2.6.0",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numpy==2.1.3",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "omegaconf==2.3.0",
    "open3d==0.19.0",
    "opencv-python==4.11.0.86",
    "packaging==25.0",
    "pandas==2.3.2",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "plotly==6.3.1",
    "plyfile==1.1.2",
    "prompt_toolkit==3.0.52",
    "psutil==7.1.0",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "Pygments==2.19.2",
    "pyparsing==3.2.5",
    "pyquaternion==0.9.9",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2025.9.18",
    "requests==2.32.5",
    "retrying==1.4.2",
    "rpds-py==0.27.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "setuptools==80.9.0",
    "six==1.17.0",
    "SQLAlchemy==2.0.43",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "timm==1.0.20",
    "tokenizers==0.22.1",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.56.2",
    "triton==3.4.0",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wandb==0.22.3",
    "wcmatch==10.1",
    "wcwidth==0.2.14",
    "Werkzeug==3.1.3",
    "widgetsnbextension==4.0.14",
    "zipp==3.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ipython>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]